added navigation, improved code modularity
