+++ frontend/src/App.js
@@ -1,61 +1,8 @@
-import logo from './logo.svg';
-import './App.css';
-import { useEffect } from "react";
-import Axios from "axios";
+import "./App.css";
+import Navigator from "./navigator";
 
-
-function App() {
-
-  // The URL to make requests to
-  const rootUrl = "https://localhost:5000";
-
-
-  const exampleRequest = async () => {
-
-    try {
-      const requestResponse = await Axios.post(
-        `${rootUrl}/users/examplepost`, // URL
-        null, // Request Body
-      );
-
-      console.log(requestResponse);
-
-    } catch(err) {
-      console.log(err);
-
-    }
-
-  }
-
-  //Use Effect: example request
-  useEffect(() => {
-    //requestResponse();
-    exampleRequest()
-  }, []);
-
-
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-
-      </header>
-    </div>
-  );
-
-
-
-}
+const App = () => {
+  return <Navigator />;
+};
 
 export default App;
