+++ frontend/src/views/index.js
@@ -1,0 +1,1 @@
+export * from "./home";
